+++ src/serena/agent.py
@@ -7,6 +7,7 @@ import multiprocessing
 import os
 import platform
 import signal
+import subprocess
 import threading
 from collections import defaultdict
 from collections.abc import Callable, Iterator, Sequence
@@ -40,6 +41,7 @@ from serena.config.serena_config import (
     ToolInclusionDefinition,
 )
 from serena.dashboard import SerenaDashboardAPI, SerenaDashboardTrayManager, SerenaDashboardViewer, open_url_in_browser
+from serena.jetbrains import jetbrains_plugin_client
 from serena.ls_manager import LanguageServerManager
 from serena.memories.memory_manager import MemoryManager
 from serena.project import Project
@@ -60,6 +62,7 @@ from serena.util.gui import system_has_usable_display
 from serena.util.inspection import iter_subclasses
 from serena.util.logging import MemoryLogHandler
 from solidlsp.ls_config import Language
+from solidlsp.util import subprocess_util
 
 if TYPE_CHECKING:
     from serena.gui_log_viewer import GuiLogViewer
@@ -1164,6 +1167,25 @@ class SerenaAgent:
         # initialize the language server in the background (if in language server mode)
         if self.get_language_backend().is_lsp():
             self.issue_task(init_language_server_manager)
+
+        def init_jetbrains_ide() -> None:
+            assert self.serena_config.jetbrains_launch_command is not None
+            try:
+                client = jetbrains_plugin_client.JetBrainsPluginClient.from_project(project, log_warning=False)
+                log.info("Found Serena JetBrains Plugin server: %s", client)
+            except jetbrains_plugin_client.ServerNotFoundError:
+                log.info("Serena JetBrains Plugin server not found for project %s", project.project_name)
+                if self.serena_config.jetbrains_launch_command:
+                    cmd = subprocess_util.convert_shell_cmd([self.serena_config.jetbrains_launch_command, project.project_root])
+                    log.info("Launching IDE with command: %s", cmd)
+                    p = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
+                    stdout, stderr = p.communicate()
+                    if p.returncode != 0:
+                        log.error(f"Failed to launch JetBrains IDE: {stderr.decode('utf-8')}")
+
+        # for JetBrains mode, search for plugin server and spawn IDE if not found and enabled
+        if self.get_language_backend().is_jetbrains():
+            self.issue_task(init_jetbrains_ide)
 
         if self._project_activation_callback is not None:
             self._project_activation_callback()